Refactor to HMT Sync, add Redis and hybrid support
Renamed package and classes from mc-data-bridge to HMT Sync, updated plugin metadata, and restructured codebase under de.jumpstone.plugins. Added Redis integration with hybrid Redis-MySQL caching, new managers for economy and playtime sync, and enhanced configuration and build profiles for Spigot, Velocity, and BungeeCord. Updated documentation and troubleshooting guides, and removed legacy proxy and test classes.
